Create README.md
Airbnb (ABNB) is an online marketplace that connects people who want to rent out their homes with people who are looking for accommodations in specific locales.Airbnb offers people an easy, relatively stress-free way to earn some income from their property

This project is an exploratory data analysis (EDA) on Airbnb booking data. The goal of the analysis is to gain insights into the patterns and trends of Airbnb bookings in the city of New York. The dataset used for the analysis contains information on the listings, reviews, and bookings of Airbnb properties in New York City from 2019.